refactor(LocalAddressBook): extract SectionHeader
