+++ src/pages/options/AddressesSettings/LocalAddressBook.tsx
@@ -3,10 +3,8 @@ import {
   Alert,
   AlertTitle,
   Box,
-  Button,
   Stack,
   Accordion,
-  AccordionSummary,
   AccordionDetails,
   CircularProgress,
   Dialog,
@@ -14,28 +12,17 @@ import {
   DialogContent,
   DialogContentText,
   DialogTitle,
-  styled,
+  Button,
 } from '@mui/material';
-import { ExpandMore as ExpandMoreIcon, Delete as DeleteIcon } from '@mui/icons-material';
 import * as murmurhash from 'murmurhash';
 
 import { useAddressBook } from '@src/shared/hooks/useAddressBook';
 
 import SectionToolbar from './SectionToolbar';
+import SectionHeader from './SectionHeader';
 import CodeMirrorTextAddresses from './CodeMirrorTextAddresses';
-import EditableTitle from './EditableTitle';
 
 import './LocalAddressBook.css';
-
-const StyledAccordionSummary = styled(AccordionSummary)(({ theme }) => ({
-  flexDirection: 'row-reverse',
-  '& .MuiAccordionSummary-expandIconWrapper': {
-    marginRight: theme.spacing(1),
-  },
-  '&.Mui-focusVisible': {
-    backgroundColor: 'transparent',
-  },
-}));
 
 interface LocalAddressBookProps {
   sectionId: string;
@@ -197,33 +184,12 @@ export default function LocalAddressBook({ sectionId }: LocalAddressBookProps) {
   return isLoaded ? (
     <>
       <Accordion defaultExpanded={true}>
-        <StyledAccordionSummary
-          expandIcon={<ExpandMoreIcon />}
-          aria-controls={`panel-${sectionId}-content`}
-          id={`panel-${sectionId}-header`}
-          title="Click to expand/collapse">
-          <Box
-            sx={{
-              display: 'flex',
-              alignItems: 'center',
-              width: '100%',
-              justifyContent: 'space-between',
-            }}>
-            <Box sx={{ display: 'flex', alignItems: 'center' }}>
-              <EditableTitle title={title} onTitleChange={handleTitleChange} />
-            </Box>
-            <Button
-              className="section-delete-button"
-              variant="contained"
-              onClick={handleDeleteClick}
-              startIcon={<DeleteIcon />}
-              size="small"
-              color="warning"
-              sx={{ mr: 1 }}>
-              Delete section
-            </Button>
-          </Box>
-        </StyledAccordionSummary>
+        <SectionHeader
+          sectionId={sectionId}
+          title={title}
+          onTitleChange={handleTitleChange}
+          onDelete={handleDeleteClick}
+        />
         <AccordionDetails sx={{ ml: '40px', mt: 0 }}>
           <SectionToolbar
             fetchUrl={fetchUrl}
